+++ karibu-testing-v10/kt10-testrun-vaadin18/src/test/kotlin/AllTests.kt
@@ -1,5 +1,6 @@
 import com.github.mvysny.dynatest.DynaTest
 import com.github.mvysny.kaributesting.v10.*
+import com.github.mvysny.kaributesting.v10.mock.MockVaadinHelper
 import com.vaadin.flow.component.littemplate.MockInstantiatorV18
 import java.net.URL
 import kotlin.test.expect
