+++ karibu-testing-v10/kt10-testrun-vaadin18/src/test/kotlin/LitTemplateTest.kt
@@ -1,7 +1,7 @@
 import com.github.mvysny.dynatest.DynaTest
 import com.github.mvysny.dynatest.expectThrows
 import com.github.mvysny.kaributesting.v10.MockVaadin
-import com.github.mvysny.kaributesting.v10.MockVaadinHelper
+import com.github.mvysny.kaributesting.v10.mock.MockVaadinHelper
 import com.github.mvysny.kaributesting.v10._isVisible
 import com.vaadin.flow.component.Component
 import com.vaadin.flow.component.Tag
