+++ karibu-testing-v10/kt10-testrun-vaadin18/src/test/kotlin/NpmPolymerTemplateTest.kt
@@ -4,7 +4,7 @@ import com.github.appreciated.app.layout.component.applayout.LeftLayouts
 import com.github.mvysny.dynatest.DynaTest
 import com.github.mvysny.dynatest.expectThrows
 import com.github.mvysny.kaributesting.v10.MockVaadin
-import com.github.mvysny.kaributesting.v10.MockVaadinHelper
+import com.github.mvysny.kaributesting.v10.mock.MockVaadinHelper
 import com.vaadin.flow.component.Component
 import com.vaadin.flow.component.Tag
 import com.vaadin.flow.component.dependency.JsModule
