+++ karibu-testing-v10/src/main/kotlin/com/github/mvysny/kaributesting/v10/mock/MockVaadinHelper.kt
@@ -1,0 +1,23 @@
+package com.github.mvysny.kaributesting.v10.mock
+
+import com.github.mvysny.kaributesting.v10.VaadinMeta
+import com.vaadin.flow.di.Instantiator
+import com.vaadin.flow.server.VaadinServlet
+import java.io.File
+
+public object MockVaadinHelper {
+    @JvmStatic
+    public fun mockFlowBuildInfo(servlet: VaadinServlet) {
+        // we need to skip the test at DeploymentConfigurationFactory.verifyMode otherwise
+        // testing a Vaadin 15 component module in npm mode without webpack.config.js nor flow-build-info.json would fail.
+        if (VaadinMeta.flowBuildInfo == null) {
+            // probably inside a Vaadin 15 component module. create a dummy token file so that
+            // DeploymentConfigurationFactory.verifyMode() is happy.
+            val tokenFile: File = File.createTempFile("flow-build-info", "json")
+            tokenFile.writeText("{}")
+            servlet.servletContext.setInitParameter("vaadin.frontend.token.file", tokenFile.absolutePath)
+        }
+    }
+
+    public var instantiatorFactory: (Instantiator) -> Instantiator = { MockInstantiator(it) }
+}
