+++ karibu-testing-v10/src/main/kotlin/com/github/mvysny/kaributesting/v10/mock/MockedUI.kt
@@ -1,0 +1,8 @@
+package com.github.mvysny.kaributesting.v10.mock
+
+import com.vaadin.flow.component.UI
+
+/**
+ * A simple no-op UI used by default by [MockVaadin.setup]. The class is open, in order to be extensible in user's library
+ */
+public open class MockedUI : UI()
